+++ apps/carp_mobile_sensing_app/lib/src/blocs/sensing_bloc.dart
@@ -6,7 +6,7 @@ class SensingBLoC {
   final String password = "...";
   final String userId = "user@cachet.dk";
   final String uri = "https://cans.cachet.dk:443";
-  final String testStudyId = "#9 - coverage";
+  final String testStudyId = "#16 - coverage :: using geolocator package";
 
   final Sensing sensing = Sensing();
   final StudyManager manager = LocalStudyManager();