+++ packages/carp_context_package/lib/src/air_quality/air_quality_probe.dart
@@ -13,9 +13,11 @@ class AirQualityProbe extends DatumProbe {
   /// Returns the [AirQualityDatum] based on the location of the phone.
   Future<Datum> getDatum() async {
     try {
-      //LocationDto location = await locationManager.getCurrentLocation();
-      //geolocator.Position location = await geolocator.getCurrentPosition();
-      location.LocationData loc = await locationProvider.getLocation();
+      print('$runtimeType - getDatum() - 1');
+      //LocationDto loc = await locationManager.getCurrentLocation();
+      //location.LocationData loc = await locationProvider.getLocation();
+      Position loc = await Geolocator.getCurrentPosition();
+      print('$runtimeType - getDatum() - 2 - $loc');
 
       if (loc != null) {
         AirQualityData q = await _waqi.feedFromGeoLocation(loc.latitude, loc.longitude);